+++ src/utils/generateFileHelpers.ts
@@ -1,0 +1,5 @@
+export const generateRandomFileName = (originalFileName: string) => {
+  const randomNumber = Math.floor(Math.random() * 1000000);
+  const fileExtension = originalFileName.split('.').pop();
+  return `${randomNumber}.${fileExtension}`;
+};
